+++ NEWS.md
@@ -1,3 +1,26 @@
+# INBOmd 0.4.6 (2018-07-19)
+
+- slides gains a logo argument
+- add report templates with Flemish corporate identity in Dutch and English
+- update Flanders Art font
+
+# INBOmd 0.4.5 (2018-06-12)
+
+- `inbo_rapport()` gains a `tocdepth` argument
+- updated installation instructions
+
+# INBOmd 0.4.4 (2018-03-12)
+
+- INBO theme for English presentation
+
+# INBOmd 0.4.3 (2018-12-18)
+
+- add `inbo_poster()`
+- add `inbo_report_css()`
+- add report template for gitbook
+- import report template for pdf
+- create addins for RStudio
+
 # INBOmd 0.4.2 (2018-04-10)
 
 - `inbo_rapport()` gains a `flandersfont` argument